refactor(activities): decompose activity_feed_card into reusable components
Split monolithic activity_feed_card.dart into focused, single-responsibility widgets:

New widget files:
- activity_feed_card_header.dart: Author, timestamp, action menu
- activity_feed_card_stats_row.dart: Views, likes, comments counts with icons
- activity_feed_card_badges.dart: Category, difficulty, distance badges
- activity_feed_card_map_thumbnail.dart: Inline map preview with tap navigation
- activity_feed- activity_feed- activity_feed- activity_feed- activity_feedat- activity_feed- activity_feed- activity_feed- activity_feed- actstr- activity_feed- activity_feed- activity_feed- activity_feed- activity_feat- activity_feed- activity_feed- actito activity_feed_formatters module
- Preserved all original functionality and styling

Benefits:
- Improved code maintainability (each widget <150 LOC)
- Better widget reusability across screens
- Easier testing of isolated formatting logic
- Clearer separation of concerns
- Reduced cognitive load for future refactors
